requirements.txt
alembic==1.10.3
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
incremental==22.10.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
msgpack==1.0.5
oauthlib==3.2.2
psycopg==3.1.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools-trial==0.6.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==1.4.47
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==2.0.3
uvicorn==0.21.1
zipp==3.15.0
zope.interface==6.0
